Extract the login function to a sepated file
